[Refactor] 파일 관련 수정
이미지 다운로드를 위한 수정
