+++ src/main/java/com/nhnacademy/bookpubshop/file/repository/FileRepository.java
@@ -11,4 +11,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
  **/
 public interface FileRepository extends JpaRepository<File, Long> {
     File findByFilePath(String path);
+
+    boolean existsByFilePath(String path);
 }
